Refactor language specific parts
