+++ pkg/codegen/python/gen_program_test.go
@@ -4,11 +4,31 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/pulumi/pulumi/pkg/v3/codegen/internal/test"
+	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
+	"github.com/pulumi/pulumi/sdk/v3/python"
 )
 
 var testdataPath = filepath.Join("..", "internal", "test", "testdata")
 
 func TestGenerateProgram(t *testing.T) {
-	test.TestProgramCodegen(t, "python", GenerateProgram)
+	test.TestProgramCodegen(t, test.ProgramLangConfig{
+		Language:   "python",
+		Extension:  "py",
+		OutputFile: "__main__.py",
+		Check:      pythonCheck,
+		GenProgram: GenerateProgram,
+	})
+}
+
+func pythonCheck(t *testing.T, path string) {
+	ex, _, err := python.CommandPath()
+	assert.NoError(t, err)
+	name := filepath.Base(path)
+	dir := filepath.Dir(path)
+	err = integration.RunCommand(t, "python syntax check",
+		[]string{ex, "-m", "py_compile", name}, dir, &integration.ProgramTestOptions{})
+	assert.NoError(t, err)
 }
